Applied patch from Terry Smith for JETTY-126 handle content > Integer.MAX_VALUE

+++ modules/jetty/src/main/java/org/mortbay/jetty/HttpConnection.java
@@ -721,7 +721,7 @@ public class HttpConnection
          * 
          * @see org.mortbay.jetty.HttpParser.EventHandler#messageComplete(int)
          */
-        public void messageComplete(int contextLength) throws IOException
+        public void messageComplete(long contextLength) throws IOException
         {
         }
